+++ internal/ceb/child_exec.go
@@ -29,11 +29,9 @@ func (ceb *CEB) initChildCmd(ctx context.Context, cfg *config) error {
 	// goroutine that will eventually run them.
 	childCmdCh := make(chan *exec.Cmd, 5)
 	doneCh := make(chan error, 1)
-	readyCh := make(chan struct{})
-	go ceb.watchChildCmd(ctx, readyCh, childCmdCh, doneCh)
+	go ceb.watchChildCmd(ctx, childCmdCh, doneCh)
 	ceb.childCmdCh = childCmdCh
 	ceb.childDoneCh = doneCh
-	ceb.childReadyCh = readyCh
 
 	return nil
 }
@@ -41,10 +39,7 @@ func (ceb *CEB) initChildCmd(ctx context.Context, cfg *config) error {
 // markChildCmdReady will allow watchChildCmd to begin executing commands.
 // This should be called once and should not be called concurrently.
 func (ceb *CEB) markChildCmdReady() {
-	if ceb.childReadyCh != nil {
-		close(ceb.childReadyCh)
-		ceb.childReadyCh = nil
-	}
+	ceb.setState(&ceb.stateChildReady, true)
 }
 
 // watchChildCmd should be started in a goroutine. This will run the child
@@ -52,7 +47,6 @@ func (ceb *CEB) markChildCmdReady() {
 // at a time.
 func (ceb *CEB) watchChildCmd(
 	ctx context.Context,
-	readyCh <-chan struct{},
 	cmdCh <-chan *exec.Cmd,
 	doneCh chan<- error,
 ) {
@@ -60,15 +54,19 @@ func (ceb *CEB) watchChildCmd(
 	// detect this and also exit.
 	defer close(doneCh)
 
-	log := ceb.logger.Named("watchChildCmd")
+	log := ceb.logger.Named("child")
 
-	// We need to wait for readyCh. readyCh is closed by ceb/init.go
+	// We need to wait for stateChildReady. This is set by ceb/init.go
 	// or ceb/config.go when we're ready to begin processing. We have to do
 	// this because we want to try to connect to the server to get our initial
 	// config values before executing our child command. But if that fails, we
 	// execute the child command anyways.
-	log.Debug("waiting for child watch loop readyCh to close")
-	<-readyCh
+	log.Debug("waiting for stateChildReady to flip to true")
+	if ceb.waitState(&ceb.stateChildReady, true) {
+		// Early exit request
+		log.Warn("exit state received before child was ready to start")
+		return
+	}
 
 	log.Debug("starting child command watch loop")
 	var currentCh <-chan error
@@ -81,19 +79,19 @@ func (ceb *CEB) watchChildCmd(
 			// If we have a child, we need to exit that first.
 			if currentCh != nil {
 				log.Info("terminating current child process")
-				err := ceb.termChildCmd(log, currentCmd, currentCh, false)
+				err := ceb.termChildCmd(log, currentCmd, currentCh, false, false)
 				log.Info("child process termination result", "err", err)
 			}
 
 			return
 
 		case cmd := <-cmdCh:
-			log.Info("child command received")
+			log.Debug("child command received")
 
 			// If we have an existing process, we need to exit that first.
 			if currentCh != nil {
-				log.Info("terminating current child process")
-				err := ceb.termChildCmd(log, currentCmd, currentCh, false)
+				log.Info("terminating current child process for restart")
+				err := ceb.termChildCmd(log, currentCmd, currentCh, false, false)
 				if err != nil {
 					// In the event terminating the child fails, we exit
 					// the whole CEB because we can't be sure we're not just
@@ -105,7 +103,7 @@ func (ceb *CEB) watchChildCmd(
 			}
 
 			// Drain the cmdCh. During graceful termination as well as initial
-			// readyCh waiting, we may accumulate a small buffer of command
+			// ready state waiting, we may accumulate a small buffer of command
 			// changes. Let's drain that.
 		CMD_DRAIN:
 			for {
@@ -117,7 +115,6 @@ func (ceb *CEB) watchChildCmd(
 			}
 
 			// Start our new command.
-			log.Info("starting new child process")
 			currentCh = ceb.startChildCmd(log, cmd)
 			currentCmd = cmd
 
@@ -180,12 +177,13 @@ func (ceb *CEB) termChildCmd(
 	cmd *exec.Cmd,
 	childErrCh <-chan error, // error channel from startChildCmd
 	force bool,
+	returnExitErr bool, // if true, doesn't treat exec.ExitError as nil
 ) error {
 	log = log.With("pid", cmd.Process.Pid)
 
 	// If we're not forcing, try a SIGTERM first.
 	if !force {
-		log.Info("sending SIGTERM")
+		log.Debug("sending SIGTERM")
 		if err := cmd.Process.Signal(syscall.SIGTERM); err != nil {
 			log.Warn("error sending SIGTERM, will proceed to SIGKILL", "err", err)
 		} else {
@@ -194,7 +192,7 @@ func (ceb *CEB) termChildCmd(
 			case err := <-childErrCh:
 				// If we got an exit error then everything worked propertly so
 				// we just set error to nil.
-				if _, ok := err.(*exec.ExitError); ok {
+				if _, ok := err.(*exec.ExitError); ok && !returnExitErr {
 					err = nil
 				}
 
@@ -212,7 +210,7 @@ func (ceb *CEB) termChildCmd(
 	// SIGKILL. We send the signal to the negative value of the pid so
 	// that it goes to the entire process group, therefore killing all
 	// grandchildren of our child process as well.
-	log.Info("sending SIGKILL")
+	log.Debug("sending SIGKILL")
 	if err := syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL); err != nil {
 		log.Warn("error sending SIGKILL", "err", err)
 		return err
@@ -221,12 +219,12 @@ func (ceb *CEB) termChildCmd(
 	// Wait for the process to die
 	log.Debug("waiting for child process to end")
 	err := <-childErrCh
-	if _, ok := err.(*exec.ExitError); ok {
+	if _, ok := err.(*exec.ExitError); ok && !returnExitErr {
 		// If we got an exit error then everything worked propertly so
 		// we just set error to nil.
 		err = nil
 	}
-	log.Info("child process exited", "wait_err", err)
+	log.Debug("child process exited", "wait_err", err)
 	return err
 }