+++ internal/serverinstall/docker.go
@@ -14,16 +14,30 @@ import (
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/docker/go-connections/nat"
+
 	"github.com/hashicorp/waypoint-plugin-sdk/terminal"
 	"github.com/hashicorp/waypoint/internal/clicontext"
+	"github.com/hashicorp/waypoint/internal/pkg/flag"
 	pb "github.com/hashicorp/waypoint/internal/server/gen"
 	"github.com/hashicorp/waypoint/internal/serverconfig"
 )
 
-func InstallDocker(
-	ctx context.Context, ui terminal.UI, scfg *Config) (
-	*clicontext.Config, *pb.ServerConfig_AdvertiseAddr, string, error,
-) {
+type DockerInstaller struct {
+	config dockerConfig
+}
+
+type dockerConfig struct {
+	serverImage string `hcl:"server_image,optional"`
+}
+
+// Install is a method of DockerInstaller and implements the Installer interface to
+// create a waypoint-server as a Docker container
+func (i *DockerInstaller) Install(
+	ctx context.Context,
+	opts *InstallOpts,
+) (*InstallResults, error) {
+	ui := opts.UI
+
 	sg := ui.StepGroup()
 	defer sg.Wait()
 
@@ -32,7 +46,7 @@ func InstallDocker(
 
 	cli, err := client.NewClientWithOpts(client.FromEnv)
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 	cli.NegotiateAPIVersion(ctx)
 
@@ -45,7 +59,7 @@ func InstallDocker(
 		}),
 	})
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	grpcPort := "9701"
@@ -74,14 +88,17 @@ func InstallDocker(
 		s.Update("Detected existing Waypoint server.")
 		s.Status(terminal.StatusWarn)
 		s.Done()
-		return &clicfg, &addr, "", nil
+		return &InstallResults{
+			Context:       &clicfg,
+			AdvertiseAddr: &addr,
+		}, nil
 	}
 
-	s.Update("Checking for Docker image: %s", scfg.ServerImage)
+	s.Update("Checking for Docker image: %s", i.config.serverImage)
 
-	imageRef, err := reference.ParseNormalizedNamed(scfg.ServerImage)
+	imageRef, err := reference.ParseNormalizedNamed(i.config.serverImage)
 	if err != nil {
-		return nil, nil, "", fmt.Errorf("Error parsing Docker image: %s", err)
+		return nil, fmt.Errorf("Error parsing Docker image: %s", err)
 	}
 
 	imageList, err := cli.ImageList(ctx, types.ImageListOptions{
@@ -91,21 +108,21 @@ func InstallDocker(
 		}),
 	})
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	if len(imageList) == 0 {
-		s.Update("Pulling image: %s", scfg.ServerImage)
+		s.Update("Pulling image: %s", i.config.serverImage)
 
 		resp, err := cli.ImagePull(ctx, reference.FamiliarString(imageRef), types.ImagePullOptions{})
 		if err != nil {
-			return nil, nil, "", err
+			return nil, err
 		}
 		defer resp.Close()
 
 		stdout, _, err := ui.OutputWriters()
 		if err != nil {
-			return nil, nil, "", err
+			return nil, err
 		}
 
 		var termFd uintptr
@@ -115,7 +132,7 @@ func InstallDocker(
 
 		err = jsonmessage.DisplayJSONMessagesStream(resp, s.TermOutput(), termFd, true, nil)
 		if err != nil {
-			return nil, nil, "", fmt.Errorf("unable to stream pull logs to the terminal: %s", err)
+			return nil, fmt.Errorf("unable to stream pull logs to the terminal: %s", err)
 		}
 
 		s.Done()
@@ -128,7 +145,7 @@ func InstallDocker(
 		Filters: filters.NewArgs(filters.Arg("label", "use=waypoint")),
 	})
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	if len(nets) == 0 {
@@ -143,19 +160,19 @@ func InstallDocker(
 		})
 
 		if err != nil {
-			return nil, nil, "", err
+			return nil, err
 		}
 
 	}
 
 	npGRPC, err := nat.NewPort("tcp", grpcPort)
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	npHTTP, err := nat.NewPort("tcp", httpPort)
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	s.Update("Installing Waypoint server to docker")
@@ -166,7 +183,7 @@ func InstallDocker(
 		AttachStdin:  true,
 		OpenStdin:    true,
 		StdinOnce:    true,
-		Image:        scfg.ServerImage,
+		Image:        i.config.serverImage,
 		ExposedPorts: nat.PortSet{npGRPC: struct{}{}, npHTTP: struct{}{}},
 		Env:          []string{"PORT=" + grpcPort},
 		Cmd:          []string{"server", "run", "-accept-tos", "-vvv", "-db=/data/data.db", "-listen-grpc=0.0.0.0:9701", "-listen-http=0.0.0.0:9702"},
@@ -200,12 +217,12 @@ func InstallDocker(
 
 	cr, err := cli.ContainerCreate(ctx, &cfg, &hostconfig, &netconfig, "waypoint-server")
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	err = cli.ContainerStart(ctx, cr.ID, types.ContainerStartOptions{})
 	if err != nil {
-		return nil, nil, "", err
+		return nil, err
 	}
 
 	// KLUDGE: There isn't a way to find out if the container is up or not,
@@ -216,5 +233,89 @@ func InstallDocker(
 	s = sg.Add("Server container started!")
 	s.Done()
 
-	return &clicfg, &addr, httpAddr, nil
+	return &InstallResults{
+		Context:       &clicfg,
+		AdvertiseAddr: &addr,
+		HTTPAddr:      httpAddr,
+	}, nil
+}
+
+// InstallRunner implements Installer by starting a single runner container.
+func (i *DockerInstaller) InstallRunner(
+	ctx context.Context,
+	opts *InstallRunnerOpts,
+) error {
+	ui := opts.UI
+
+	sg := ui.StepGroup()
+	defer sg.Wait()
+
+	s := sg.Add("Initializing Docker client...")
+	defer func() { s.Abort() }()
+
+	cli, err := client.NewClientWithOpts(client.FromEnv)
+	if err != nil {
+		return err
+	}
+	cli.NegotiateAPIVersion(ctx)
+
+	s.Update("Checking for an existing runner...")
+	containers, err := cli.ContainerList(ctx, types.ContainerListOptions{
+		Filters: filters.NewArgs(filters.KeyValuePair{
+			Key:   "label",
+			Value: "waypoint-type=runner",
+		}),
+	})
+	if err != nil {
+		return err
+	}
+	if len(containers) > 0 {
+		s.Update("Detected existing Waypoint runner.")
+		s.Status(terminal.StatusWarn)
+		s.Done()
+		return nil
+	}
+
+	// The key thing in the container creation below is that the environment
+	// variables are set to the advertised address env vars which will
+	// allow our runner to connect.
+	cr, err := cli.ContainerCreate(ctx, &container.Config{
+		AttachStdout: true,
+		AttachStderr: true,
+		AttachStdin:  true,
+		OpenStdin:    true,
+		StdinOnce:    true,
+		Image:        i.config.serverImage,
+		Env:          opts.AdvertiseClient.Env(),
+		Cmd:          []string{"runner", "agent", "-vvv"},
+		Labels: map[string]string{
+			"waypoint-type": "runner",
+		},
+	}, nil, &network.NetworkingConfig{
+		EndpointsConfig: map[string]*network.EndpointSettings{
+			"waypoint": {},
+		},
+	}, "waypoint-runner")
+	if err != nil {
+		return err
+	}
+
+	err = cli.ContainerStart(ctx, cr.ID, types.ContainerStartOptions{})
+	if err != nil {
+		return err
+	}
+
+	s.Update("Waypoint runner installed and started!")
+	s.Done()
+
+	return nil
+}
+
+func (i *DockerInstaller) InstallFlags(set *flag.Set) {
+	set.StringVar(&flag.StringVar{
+		Name:    "docker-server-image",
+		Target:  &i.config.serverImage,
+		Usage:   "Docker image for the Waypoint server.",
+		Default: "hashicorp/waypoint:latest",
+	})
 }