+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## unreleased
+## 0.2.0 (December 10, 2020)
 
 FEATURES:
 
+* **Application config syncing with Kubernetes (ConfigMaps), Vault, Consul, and AWS SSM**;
+Automatically sync environment variable values with remote sources and restart your
+application when those values change. [GH-810]
 * **Access to Artifact, Deploy Metadata**: `registry` and `deploy` configuration can use 
 `artifact.*` variable syntax to access metadata from the results of those stages. 
 The `release` configuration can use `artifact.*` and `deploy.*` to access metadata.
@@ -12,14 +15,36 @@ available to your Waypoint configuration.
 * **`path` Variables**: you can now use `path.project`, `path.app`, and `path.pwd` as
 variables in your Waypoint file to specify paths as relative to the project (waypoint.hcl
 file), app, or pwd of the CLI invocation.
+* **Server snapshot/restore**: you can now use the CLI or API to take and restore online
+snapshots. Restoring snapshot data requires a server restart but the restore operation 
+can be staged online. [GH-870]
 
 IMPROVEMENTS:
 
+* cli/logs: entrypoint logs can now be seen alongside app logs and are colored differently [GH-855]
+* contrib/serverinstall: Automate setup of kind+k8s with metallb [GH-845]
 * core: application config changes (i.e. `waypoint config set`) will now restart running applications [GH-791]
+* core: add more descriptive text to include app name in `waypoint destroy` [GH-807]
+* core: add better error messaging when prefix is missing from the `-raw` flag in `waypoint config set` [GH-815]
+* core: align -raw flag to behave like -json flag with waypoint config set [GH-828]
+* core: `waypoint.hcl` can be named `waypoint.hcl.json` and use JSON syntax [GH-867]
+* install: Update flags used on server install per-platform [GH-882]
 * install/k8s: support for OpenShift [GH-715]
+* internal/server: Block on instance deployment becoming available [GH-881]
+* plugin/aws-ecr: environment variables to be used instead of 'region' property for aws-ecr registry [GH-841]
+* plugin/google-cloud-run: allow images from Google Artifact Registry [GH-804]
+* plugin/google-cloud-run: added service account name field [GH-850]
 * server: APIs for Waypoint database snapshot/restore [GH-723]
+* website: many minor improvements were made in our plugin documentation section for this release
 
 BUG FIXES:
+
+* core: force killed `waypoint exec` sessions won't leave the remote process running [GH-827]
+* core: waypoint exec with no TTY won't hang open until a ctrl-c [GH-830]
+* cli: server config-set doesn't require a Waypoint configuration file. [GH-819]
+* cli/token: fix issue where tokens could be cut off on narrow terminals [GH-885]
+* plugin/aws-ecs: task_role_name is optional [GH-824]
+
 
 ## 0.1.5 (November 09, 2020)